+++ src/Either.ts
@@ -6,7 +6,8 @@ import { Either } from "fp-ts/Either"
 import * as E from "fp-ts/Either"
 
 /**
- * Unwrap the value from within an `Either`, throwing if `Left`.
+ * Unwrap the value from within an `Either`, throwing the inner value of `Left`
+ * if `Left`.
  *
  * @example
  * import { unsafeUnwrap } from 'fp-ts-std/Either';
@@ -18,13 +19,14 @@ import * as E from "fp-ts/Either"
  */
 export const unsafeUnwrap = <A>(x: Either<unknown, A>): A => {
   // eslint-disable-next-line functional/no-conditional-statement, functional/no-throw-statement
-  if (E.isLeft(x)) throw "Unsafe attempt to unwrap Either failed"
+  if (E.isLeft(x)) throw x.left
 
   return x.right
 }
 
 /**
- * Unwrap the value from within an `Either`, throwing if `Right`.
+ * Unwrap the value from within an `Either`, throwing the inner value of `Right`
+ * if `Right`.
  *
  * @example
  * import { unsafeUnwrapLeft } from 'fp-ts-std/Either';
@@ -36,7 +38,7 @@ export const unsafeUnwrap = <A>(x: Either<unknown, A>): A => {
  */
 export const unsafeUnwrapLeft = <E>(x: Either<E, unknown>): E => {
   // eslint-disable-next-line functional/no-conditional-statement, functional/no-throw-statement
-  if (E.isRight(x)) throw "Unsafe attempt to unwrap Either failed"
+  if (E.isRight(x)) throw x.right
 
   return x.left
 }